% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_constant_metadata_dataset.R
\name{add_constant_metadata_dataset}
\alias{add_constant_metadata_dataset}
\title{add_constant_metadata_dataset}
\usage{
add_constant_metadata_dataset(metadata_list)
}
\arguments{
\item{metadata_list}{list: List returned by flatten_md_metadata()}
}
\value{
list
}
\description{
Add metadata that have constant values across records
(not pulled from Microdata API)
Currently adds values for the following fields:
field_wbddh_languages_supported ('English')
field_wbddh_data_type ('Microdata')
field_contact_email ("microdata at worldbank.org")
field_ddh_harvest_src ("Microdata")
}
