% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_country}
\alias{add_country}
\title{add_country}
\usage{
add_country(metadata_in, iso3_lkup)
}
\arguments{
\item{iso3_lkup}{data.frame: ddhconnect::get_iso3()}
}
\value{
list
}
\description{
Add country name based on iso3 code
Use field_wbddh_reference_id to
- Extract ISO 3 code
- Map to country name using ddhconnect::get_iso3()
Consider using tryCatch for multiple ISO3s
}
