% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data_classification.R
\name{add_data_classification}
\alias{add_data_classification}
\title{add_data_classification}
\usage{
add_data_classification(metadata_list, md_internal_id, master)
}
\arguments{
\item{metadata_list}{list: Flattened list of metadata from the Microdata API}

\item{md_internal_id}{character: Microdata internal ID}

\item{master}{data.frame: Output of mdlibtoddh::get_ddh_records_status()}
}
\value{
list
}
\description{
add_data_classification
}
