% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_dataset.R
\name{add_new_dataset}
\alias{add_new_dataset}
\title{add_new_dataset()}
\usage{
add_new_dataset(md_internal_id, md_token, ddh_credentials, master, root_url)
}
\arguments{
\item{md_internal_id}{character: Microdata internal ID of the dataset to be added}

\item{md_token}{character: Microdata API authentication token}

\item{ddh_credentials}{list: DDH API authentication token and cookie}

\item{master}{dataframe: Master lookup table}

\item{root_url}{character: Root URL to use for the API (Staging or Production)}
}
\value{
character
}
\description{
Add a full microdata record to DDH (metadata + resources)
}
