% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_search_tags.R
\name{add_search_tags}
\alias{add_search_tags}
\title{add_search_tags}
\usage{
add_search_tags(metadata_list, id, token, ...)
}
\arguments{
\item{metadata_list}{list: List of metadata pulled from Microdata API}

\item{id}{character: Microdata study internal ID}

\item{token}{character: Microdata API authentication token}

\item{...}{...: Extra paramter to be passed to add_search_tags (See get_md_dictionary() limit parameter)}
}
\description{
Add search tags to the list of variable to be added to DDH
}
