% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ddh_tid_lovs.R
\docType{data}
\name{ddh_tid_lovs}
\alias{ddh_tid_lovs}
\title{DDH values to DDH IDs mapping}
\format{
A list of 6 named vectors.
}
\usage{
ddh_tid_lovs
}
\description{
List of named vectors to facilitate the mapping of DDH list of values (LOVs) to DDH LOV IDs
}
\examples{
ddh_tid_lovs
}
\keyword{datasets}
