% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_md_metadata.R
\name{extract_md_metadata}
\alias{extract_md_metadata}
\title{extract_md_metadata}
\usage{
extract_md_metadata(metadata_in,
  metadata_out = mdlibtoddh::md_placeholder,
  lookup = mdlibtoddh::lookup)
}
\arguments{
\item{metadata_in}{list: The output of mdlibconnect::get_survey_metadata}

\item{metadata_out}{list: Package object: mdlibtoddh::md_placeholder}

\item{lookup}{data.frame: Package object: mdlibtoddh::lookup}
}
\value{
list
}
\description{
Extract specific metadata from the Microdata API JSON response
}
