% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_md_metadata.R
\name{flatten_md_metadata}
\alias{flatten_md_metadata}
\title{flatten_md_metadata}
\usage{
flatten_md_metadata(metadata_list,
  multiple_fields = c("field_wbddh_primary_investigator",
  "field_wbddh_authority_name", "field_wbddh_publisher_name"),
  email_fields = c("field_contact_email"),
  name_fields = c("field_wbddh_country", "field_wbddh_region",
  "field_wbddh_data_collector"))
}
\arguments{
\item{metadata_list}{list: Output of mdlibtoddh::extract_md_metadata}

\item{multiple_fields}{character: JSON array fields in Microdata}

\item{email_fields}{character: email fiels}

\item{name_fields}{character: Name fields}
}
\value{
list
}
\description{
Flatten list
}
