% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ddh_datasets_list.R
\name{get_ddh_datasets_list}
\alias{get_ddh_datasets_list}
\title{get_ddh_datasets_list}
\usage{
get_ddh_datasets_list()
}
\value{
data.frame
}
\description{
Return the full list of microdata datasets in DDH with the following information: node IDs, UUIDs, reference IDs (microdata), data classification, date of last update.
}
