% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ddh_records_status.R
\name{get_ddh_records_status}
\alias{get_ddh_records_status}
\title{get_ddh_records_status}
\usage{
get_ddh_records_status(mdlib_token)
}
\arguments{
\item{mdlib_token}{character: Microdatalib API authentication token}
}
\value{
data frame
}
\description{
Compare DDH and Microdata library records
}
