% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_md_dictionary.R
\name{get_md_dictionary}
\alias{get_md_dictionary}
\title{get_md_dictionary}
\usage{
get_md_dictionary(id, token, limit = 10000)
}
\arguments{
\item{id}{character: survey unique id (internal id)}

\item{token}{character: Microdata API authentication token}

\item{limit}{numeric: Maximum number of variables to be returned}
}
\value{
character vector
}
\description{
return a list of survey variables and list of values as a single, semi-colon separated string (required format to be passed to the DDH API)
}
