% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_md_metadata.R
\name{get_md_metadata}
\alias{get_md_metadata}
\title{get_md_metadata}
\usage{
get_md_metadata(id, token)
}
\arguments{
\item{id}{character: Microdata internal survey id}

\item{token}{character: Microdata API authentication token}
}
\value{
list
}
\description{
Fetch survey metadata from Microdata internal API
}
