% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_md_public_datasets_list.R
\name{get_md_public_datasets_list}
\alias{get_md_public_datasets_list}
\title{get_md_public_datasets_list}
\usage{
get_md_public_datasets_list(token)
}
\arguments{
\item{token}{character: Microdatalib API authentication token}
}
\value{
data.frame
}
\description{
get_md_public_datasets_list
}
