% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_array_metadata.R
\name{handle_array_metadata}
\alias{handle_array_metadata}
\title{handle_array_metadata}
\usage{
handle_array_metadata(metadata_in,
  metadata_out = mdlibtoddh::md_placeholder,
  lookup = mdlibtoddh::lookup)
}
\arguments{
\item{metadata_in}{list: The output of mdlibconnect::get_survey_metadata}

\item{metadata_out}{list: Package object: mdlibtoddh::md_placeholder}

\item{lookup}{data.frame: Package object: mdlibtoddh::lookup}
}
\value{
list
}
\description{
Handle list metadata
}
