% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-json_template_attach.R
\docType{data}
\name{json_template_attach}
\alias{json_template_attach}
\title{JSON template to attach resources to datasets}
\format{A list to be turned into a JSON object}
\usage{
json_template_attach
}
\description{
This is a list to be used as a template to generate a JSON object via jsonlite::toJSON.
}
\examples{
json_template_attach
}
\keyword{datasets}
