% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-json_template_dataset.R
\docType{data}
\name{json_template_dataset}
\alias{json_template_dataset}
\title{JSON template to create a dataset}
\format{
A list to be turned into a JSON object
}
\usage{
json_template_dataset
}
\description{
This is a list to be used as a template to generate a JSON object via jsonlite::toJSON.
}
\examples{
json_template_dataset
}
\keyword{datasets}
