% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-json_template_resource.R
\docType{data}
\name{json_template_resource}
\alias{json_template_resource}
\title{JSON template to create a resource}
\format{A list to be turned into a JSON object}
\usage{
json_template_resource
}
\description{
This is a list to be used as a template to generate a JSON object via jsonlite::toJSON.
}
\examples{
json_template_resource
}
\keyword{datasets}
