% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_md_metadata.R
\name{map_md_metadata}
\alias{map_md_metadata}
\title{map_md_metadata}
\usage{
map_md_metadata(metadata_list, lkup_tids)
}
\arguments{
\item{metadata_list}{list: Named list containing all values extracted from the Microdata API response}

\item{lkup_tids}{named vector}
}
\value{
list
}
\description{
map_md_metadata
}
