% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-md_ddh_lovs.R
\docType{data}
\name{md_ddh_lovs}
\alias{md_ddh_lovs}
\title{Microdata to DDH values mapping}
\format{
A list of 6 named vectors.
}
\usage{
md_ddh_lovs
}
\description{
List of named vectors to facilitate the mapping of Microdata values to DDH controlled vocabulary
}
\examples{
md_ddh_lovs
}
\keyword{datasets}
