% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-md_placeholder.R
\docType{data}
\name{md_placeholder}
\alias{md_placeholder}
\title{Microdata placeholder}
\format{
An empty list of length 84.
}
\usage{
md_placeholder
}
\description{
This is an named empty list to hold metadata from the Microdata API. The names of each element of the list are DDH machine names.
}
\examples{
md_placeholder
}
\keyword{datasets}
