% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_blank_values.R
\name{pass_blank_values}
\alias{pass_blank_values}
\title{pass_blank_values()}
\usage{
pass_blank_values(
  node_id,
  dataset_metadata,
  metadata_list,
  root_url,
  credentials
)
}
\arguments{
\item{node_id}{character: DDH node ID of the dataset to be updated}

\item{dataset_metadata}{character: The dataset metadata, returned from ddhconnect::get_metadata()}

\item{metadata_list}{list: List of metatadata values}

\item{root_url}{character: Root URL to use for the API (Staging or Production)}

\item{credentials}{list: DDH API authentication token and cookie}
}
\value{
list
}
\description{
Pass blank values for values which are not present in updated version of dataset
}
