% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_existing_dataset.R
\name{update_existing_dataset}
\alias{update_existing_dataset}
\title{update_existing_dataset()}
\usage{
update_existing_dataset(md_internal_id, md_token, ddh_credentials, master,
  root_url)
}
\arguments{
\item{md_internal_id}{character: Microdata internal ID of the dataset to be added}

\item{md_token}{character: Microdata API authentication token}

\item{ddh_credentials}{list: DDH API authentication token and cookie}

\item{master}{dataframe: Master lookup table}

\item{root_url}{character: Root URL to use for the API (Staging or Production)}
}
\value{
character
}
\description{
Update a full microdata record in DDH (metadata + resources)
}
