% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_md_metadata.R
\name{format_md_metadata}
\alias{format_md_metadata}
\title{format_md_metadata}
\usage{
format_md_metadata(metadata_in, date_fields = c("field_wbddh_modified_date",
  "field_wbddh_release_date"))
}
\arguments{
\item{metadata_in}{list: flattened list of metadata}
}
\value{
list
}
\description{
Format specific variable to fit Microdata or DDH requirements
}
