% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ddh_datasets_list.R
\name{get_ddh_datasets_list}
\alias{get_ddh_datasets_list}
\title{get_ddh_datasets_list}
\usage{
get_ddh_datasets_list(root_url = dkanr::get_url(), credentials = list(cookie
  = dkanr::get_cookie(), token = dkanr::get_token()))
}
\arguments{
\item{root_url}{character: API root URL}

\item{credentials}{list: object returned by the get_credentials() function}
}
\value{
data.frame
}
\description{
Return the full list of microdata datasets in DDH with the following information: node IDs, UUIDs, reference IDs (microdata), data classification, date of last update.
}
