% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ddh_records_status.R
\name{get_ddh_records_status}
\alias{get_ddh_records_status}
\title{get_ddh_records_status}
\usage{
get_ddh_records_status(mdlib_token,
  root_url = ddhconnect:::production_root_url)
}
\arguments{
\item{mdlib_token}{character: Microdatalib API authentication token}

\item{root_url}{character: API root URL}
}
\value{
data frame
}
\description{
Compare DDH and Microdata library records
}
