% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_md_to_ddh.R
\name{map_md_to_ddh}
\alias{map_md_to_ddh}
\title{map_md_to_ddh}
\usage{
map_md_to_ddh(metadata_list, lkup_tids)
}
\arguments{
\item{metadata_list}{list: List of Microdata metadata (output from get_md_metadata())}

\item{lkup_tids}{named vector}
}
\value{
list
}
\description{
Map response from Microdata to DDH variables and LOVs
}
